move builder class into main package, deprecate interface
